requirements.txt
asgiref==3.3.1
coverage==5.5
Django==3.1.7
djangorestframework==3.12.2
Pillow==8.1.2
pkg-resources==0.0.0
pytz==2021.1
sqlparse==0.4.1
